refactor: simplify my collection artwork form
