+++ src/app/Scenes/MyCollection/Screens/ArtworkForm/MyCollectionArtworkForm.tsx
@@ -14,27 +14,22 @@ import { LengthUnitPreference } from "__generated__/UserPrefsModelQuery.graphql"
 import LoadingModal from "app/Components/Modals/LoadingModal"
 import { updateMyUserProfile } from "app/Scenes/MyAccount/updateMyUserProfile"
 import { AddMyCollectionArtist } from "app/Scenes/MyCollection/Screens/Artist/AddMyCollectionArtist"
+import { MyCollectionArtworkStore } from "app/Scenes/MyCollection/Screens/ArtworkForm/MyCollectionArtworkStore"
+import { updateArtwork } from "app/Scenes/MyCollection/Screens/ArtworkForm/methods/uploadArtwork"
 import { ArtworkFormValues } from "app/Scenes/MyCollection/State/MyCollectionArtworkModel"
-import { deleteArtworkImage } from "app/Scenes/MyCollection/mutations/deleteArtworkImage"
-import { myCollectionCreateArtwork } from "app/Scenes/MyCollection/mutations/myCollectionCreateArtwork"
 import { myCollectionDeleteArtwork } from "app/Scenes/MyCollection/mutations/myCollectionDeleteArtwork"
-import { myCollectionUpdateArtwork } from "app/Scenes/MyCollection/mutations/myCollectionUpdateArtwork"
-import { deletedPhotos } from "app/Scenes/MyCollection/utils/deletedPhotos"
 import { Tab } from "app/Scenes/MyProfile/MyProfileHeaderMyCollectionAndSavedWorks"
 import { GlobalStore } from "app/store/GlobalStore"
 import { goBack } from "app/system/navigation/navigate"
-import { storeLocalImage } from "app/utils/LocalImageStore"
 import { useFeatureFlag } from "app/utils/hooks/useFeatureFlag"
-import { addClue, setVisualClueAsSeen } from "app/utils/hooks/useVisualClue"
 import { refreshMyCollection, refreshMyCollectionInsights } from "app/utils/refreshHelpers"
 import { FormikProvider, useFormik } from "formik"
-import { isEqual, reverse } from "lodash"
+import { isEqual } from "lodash"
 import { useEffect, useRef, useState } from "react"
 import { Alert, InteractionManager } from "react-native"
 import { useTracking } from "react-tracking"
 import { SavingArtworkModal } from "./Components/SavingArtworkModal"
 import { artworkSchema, validateArtworkSchema } from "./Form/artworkSchema"
-import { uploadPhotos } from "./MyCollectionImageUtil"
 import { MyCollectionAddPhotos } from "./Screens/MyCollectionArtworkFormAddPhotos"
 import { MyCollectionArtworkFormArtist } from "./Screens/MyCollectionArtworkFormArtist"
 import { MyCollectionArtworkFormArtwork } from "./Screens/MyCollectionArtworkFormArtwork"
@@ -101,6 +96,10 @@ export const MyCollectionArtworkForm: React.FC<MyCollectionArtworkFormProps> = (
   const preferredCurrency = GlobalStore.useAppState((state) => state.userPrefs.currency)
   const preferredMetric = GlobalStore.useAppState((state) => state.userPrefs.metric)
 
+  const { mode, onSuccess, onDelete, artwork } = MyCollectionArtworkStore.useStoreState(
+    (state) => state
+  )
+
   // we need to store the form values in a ref so that onDismiss can access their current value (prop updates are not
   // sent through the react-navigation system)
   const formValuesRef = useRef(formValues)
@@ -146,7 +145,7 @@ export const MyCollectionArtworkForm: React.FC<MyCollectionArtworkFormProps> = (
 
     try {
       // Adding tracking after a successfully adding an artwork
-      if (props.mode === "add") {
+      if (mode === "add") {
         trackEvent(
           tracks.saveCollectedArtwork(
             values.artistIds[0],
@@ -166,7 +165,7 @@ export const MyCollectionArtworkForm: React.FC<MyCollectionArtworkFormProps> = (
     }
 
     InteractionManager.runAfterInteractions(() => {
-      props.onSuccess?.()
+      onSuccess?.()
     })
   }
 
@@ -184,27 +183,26 @@ export const MyCollectionArtworkForm: React.FC<MyCollectionArtworkFormProps> = (
     validationSchema: artworkSchema,
   })
 
-  const onDelete =
-    props.mode === "edit" && props.onDelete
-      ? async () => {
-          setLoading(true)
-          trackEvent(tracks.deleteCollectedArtwork(props.artwork.internalID, props.artwork.slug))
-          try {
-            await myCollectionDeleteArtwork(props.artwork.internalID)
-            refreshMyCollection()
-            props.onDelete()
-          } catch (e) {
-            if (__DEV__) {
-              console.error(e)
-            } else {
-              captureException(e)
-            }
-            Alert.alert("An error ocurred", typeof e === "string" ? e : undefined)
-          } finally {
-            setLoading(false)
-          }
+  const handleDelete = async () => {
+    if (mode === "edit" && onDelete && artwork) {
+      setLoading(true)
+      trackEvent(tracks.deleteCollectedArtwork(artwork.internalID, artwork.slug))
+      try {
+        await myCollectionDeleteArtwork(artwork.internalID)
+        refreshMyCollection()
+        onDelete()
+      } catch (e) {
+        if (__DEV__) {
+          console.error(e)
+        } else {
+          captureException(e)
         }
-      : undefined
+        Alert.alert("An error ocurred", typeof e === "string" ? e : undefined)
+      } finally {
+        setLoading(false)
+      }
+    }
+  }
 
   const clearForm = async () => {
     const formIsDirty = !isEqual(formValuesRef.current, dirtyFormCheckValues)
@@ -231,7 +229,7 @@ export const MyCollectionArtworkForm: React.FC<MyCollectionArtworkFormProps> = (
       }
     }
 
-    if (props.mode === "edit") {
+    if (mode === "edit") {
       // Reset the form with the initial values from the artwork
       GlobalStore.actions.myCollection.artwork.updateFormValues(dirtyFormCheckValues)
     } else {
@@ -242,7 +240,7 @@ export const MyCollectionArtworkForm: React.FC<MyCollectionArtworkFormProps> = (
   const onHeaderBackButtonPress = () => {
     const currentRoute = navContainerRef.current?.getCurrentRoute()
     const isFirstScreen =
-      props.mode === "edit" || !currentRoute?.name || currentRoute?.name === "ArtworkFormArtist"
+      mode === "edit" || !currentRoute?.name || currentRoute?.name === "ArtworkFormArtist"
 
     // clear and exit the form if we're on the first screen
     if (isFirstScreen) {
@@ -265,31 +263,31 @@ export const MyCollectionArtworkForm: React.FC<MyCollectionArtworkFormProps> = (
             cardStyle: { backgroundColor: "white" },
           }}
         >
-          {props.mode === "add" && (
+          {mode === "add" && (
             <Stack.Screen
               name="ArtworkFormArtist"
               component={MyCollectionArtworkFormArtist}
-              initialParams={{ onDelete, clearForm, onHeaderBackButtonPress }}
+              initialParams={{ onDelete: handleDelete, clearForm, onHeaderBackButtonPress }}
             />
           )}
-          {props.mode === "add" && (
+          {mode === "add" && (
             <Stack.Screen
               name="ArtworkFormArtwork"
               component={MyCollectionArtworkFormArtwork}
               initialParams={{
                 onDelete,
                 clearForm,
-                mode: props.mode,
+                mode: mode,
                 onHeaderBackButtonPress,
               }}
             />
           )}
-          {!!enableCollectedArtists && props.mode === "add" && (
+          {!!enableCollectedArtists && mode === "add" && (
             <Stack.Screen
               name="AddMyCollectionArtist"
               component={AddMyCollectionArtist} // TODO: Rename this component
               initialParams={{
-                mode: props.mode,
+                mode: mode,
                 clearForm,
                 onHeaderBackButtonPress,
               }}
@@ -301,11 +299,11 @@ export const MyCollectionArtworkForm: React.FC<MyCollectionArtworkFormProps> = (
             initialParams={{
               onDelete,
               clearForm,
-              mode: props.mode,
+              mode: mode,
               onHeaderBackButtonPress,
               isSubmission: (() => {
-                if (props.mode === "edit") {
-                  return !!props.artwork.consignmentSubmission?.displayText
+                if (mode === "edit" && artwork) {
+                  return !!artwork.consignmentSubmission?.displayText
                 }
                 return false
               })(),
@@ -313,7 +311,7 @@ export const MyCollectionArtworkForm: React.FC<MyCollectionArtworkFormProps> = (
           />
           <Stack.Screen name="AddPhotos" component={MyCollectionAddPhotos} />
         </Stack.Navigator>
-        {props.mode === "add" ? (
+        {mode === "add" ? (
           <SavingArtworkModal
             testID="saving-artwork-modal"
             isVisible={loading}
@@ -327,141 +325,15 @@ export const MyCollectionArtworkForm: React.FC<MyCollectionArtworkFormProps> = (
   )
 }
 
-const Stack = createStackNavigator<ArtworkFormScreen>()
-
-export const updateArtwork = async (
-  values: ArtworkFormValues,
-  dirtyFormCheckValues: ArtworkFormValues,
-  props: MyCollectionArtworkFormProps
-) => {
-  const {
-    photos,
-    artistSearchResult,
-    customArtist,
-    pricePaidDollars,
-    pricePaidCurrency,
-    artistDisplayName,
-    ...others
-  } = values
-
-  const newPhotos = photos.filter((photo) => photo.path)
-  const externalImageUrls = await uploadPhotos(newPhotos)
-
-  let pricePaidCents
-  if (pricePaidDollars && !isNaN(Number(pricePaidDollars))) {
-    pricePaidCents = parseFloat(pricePaidDollars) * 100
-  }
-
-  if (values.attributionClass !== "LIMITED_EDITION") {
-    others.editionNumber = ""
-    others.editionSize = ""
-  }
-
-  if (props.mode === "add") {
-    let artistsData
-
-    if (artistDisplayName) {
-      artistsData = [{ displayName: artistDisplayName }]
-    } else if (customArtist) {
-      artistsData = [{ displayName: customArtist.name }]
-    } else artistsData = undefined
-
-    const response = await myCollectionCreateArtwork({
-      artistIds: artistSearchResult?.internalID ? [artistSearchResult?.internalID] : undefined,
-      artists: artistsData,
-      artworkLocation: others.artworkLocation,
-      attributionClass: others.attributionClass || undefined,
-      category: others.category,
-      date: others.date,
-      depth: others.depth,
-      editionNumber: others.editionNumber,
-      editionSize: others.editionSize,
-      externalImageUrls,
-      height: others.height,
-      isEdition: others.isEdition,
-      medium: others.medium,
-      metric: others.metric,
-      confidentialNotes: others.confidentialNotes,
-      pricePaidCents,
-      pricePaidCurrency,
-      provenance: others.provenance,
-      title: others.title,
-      width: others.width,
-    })
-
-    const artwork = response.myCollectionCreateArtwork?.artworkOrError?.artworkEdge?.node
-
-    // Store images locally
-    await Promise.all(
-      newPhotos.map(async (image, index) => {
-        const imageID = artwork?.images?.[index]?.internalID
-
-        if (!imageID) return
-
-        await storeLocalImage(imageID, {
-          path: image.path!,
-          width: image.width!,
-          height: image.height!,
-        })
-      })
-    )
-
-    const hasMarketPriceInsights =
-      response.myCollectionCreateArtwork?.artworkOrError?.artworkEdge?.node?.hasMarketPriceInsights
-
-    addArtworkMessages({ hasMarketPriceInsights, sourceTab: props.source })
-    return hasMarketPriceInsights
-  } else {
-    const response = await myCollectionUpdateArtwork({
-      artistIds: artistSearchResult?.internalID ? [artistSearchResult?.internalID] : [],
-      artworkId: props.artwork.internalID,
-      artworkLocation: others.artworkLocation,
-      attributionClass: others.attributionClass || undefined,
-      category: others.category,
-      date: others.date,
-      depth: others.depth,
-      editionNumber: others.editionNumber,
-      editionSize: others.editionSize,
-      externalImageUrls,
-      height: others.height,
-      isEdition: others.isEdition,
-      medium: others.medium,
-      metric: others.metric,
-      confidentialNotes: others.confidentialNotes,
-      pricePaidCents: pricePaidCents ?? null,
-      pricePaidCurrency,
-      provenance: others.provenance,
-      title: others.title,
-      width: others.width,
-    })
-
-    const updatedArtwork = response.myCollectionUpdateArtwork?.artworkOrError?.artwork
-
-    // Store images locally and start from the end because
-    // it's only possible to add new images at the end
-    const reversedImages = reverse([...(updatedArtwork?.images ?? [])])
-
-    await Promise.all(
-      reverse([...newPhotos]).map(async (image, index) => {
-        const imageID = reversedImages[index]?.internalID
-
-        if (!imageID) return
-
-        await storeLocalImage(imageID, {
-          path: image.path!,
-          width: image.width!,
-          height: image.height!,
-        })
-      })
-    )
-
-    // Delete images
-    const deletedImages = deletedPhotos(dirtyFormCheckValues.photos, photos)
-    for (const photo of deletedImages) {
-      await deleteArtworkImage(props.artwork.internalID, photo.id)
-    }
-  }
+export const MyCollectionArtworkFormScreen: React.FC<MyCollectionArtworkFormProps> = (props) => {
+  return (
+    <MyCollectionArtworkStore.Provider runtimeModel={props}>
+      <MyCollectionArtworkForm {...props} />
+    </MyCollectionArtworkStore.Provider>
+  )
 }
+
+const Stack = createStackNavigator<ArtworkFormScreen>()
 
 const tracks = {
   deleteCollectedArtwork: (internalID: string, slug: string): DeleteCollectedArtwork => ({
@@ -483,34 +355,4 @@ const tracks = {
     is_p1_artist: isP1Artist ?? false,
     platform: "mobile",
   }),
-}
-
-const addArtworkMessages = async ({
-  hasMarketPriceInsights,
-  sourceTab,
-}: {
-  hasMarketPriceInsights: boolean | null | undefined
-  sourceTab: Tab
-}) => {
-  setVisualClueAsSeen("AddedArtworkWithInsightsMessage_InsightsTab")
-  setVisualClueAsSeen("AddedArtworkWithInsightsMessage_MyCTab")
-  setVisualClueAsSeen("AddedArtworkWithoutInsightsMessage_InsightsTab")
-  setVisualClueAsSeen("AddedArtworkWithoutInsightsMessage_MyCTab")
-
-  if (hasMarketPriceInsights) {
-    if (sourceTab === Tab.collection) {
-      addClue("AddedArtworkWithInsightsMessage_MyCTab")
-      addClue("AddedArtworkWithInsightsVisualClueDot")
-    } else {
-      setVisualClueAsSeen("MyCollectionInsightsIncompleteMessage")
-      addClue("AddedArtworkWithInsightsMessage_InsightsTab")
-    }
-  } else {
-    if (sourceTab === Tab.collection) {
-      addClue("AddedArtworkWithoutInsightsMessage_MyCTab")
-    } else {
-      setVisualClueAsSeen("MyCollectionInsightsIncompleteMessage")
-      addClue("AddedArtworkWithoutInsightsMessage_InsightsTab")
-    }
-  }
 }
